Crate geo_offset

This crate contains algorithms to shrink and dilate various geometric shapes.

This code is a Rust port of the JS lib polygon-offset.

Example

The following example shows how to compute an inflated line.
The offset method is provided by the Offset trait which is implemented for most geo-types.

use geo_types::{Coord, Line};
use geo_offset::Offset;

let line = Line::new(
    Coord { x: 0.0, y: 0.0 },
    Coord { x: 1.0, y: 8.0 },
);

let line_with_offset = line.offset(2.0)?;

Structs

Edge

Enums

EdgeError

This enumeration contains error cases for edges manipulation.

OffsetError

If offset computing fails this error is returned.

Constants

DEFAULT_ARC_SEGMENTS

Arcs around corners are made of 5 segments by default.

Traits

Offset